storage/requirements.txt
connexion[flask,uvicorn,swagger-ui]
PyYAML
sqlalchemy
pymysql
cryptography
kafka-python